refactor: move event parser out of calendar component
