+++ src/pages/scheduler/components/SchedulerCalendar.tsx
@@ -5,12 +5,12 @@ import 'react-big-calendar/lib/sass/styles.scss';
 import { addWeeks, format, getDay, parse, set, startOfWeek } from 'date-fns';
 import { enUS } from 'date-fns/locale';
 import { Calendar, dateFnsLocalizer } from 'react-big-calendar';
-import { RRule, Weekday } from 'rrule';
 
 import { useDarkMode } from 'lib/hooks/useDarkMode';
 
 import { CalendarEvent } from 'pages/scheduler/components/Event';
 import { CalendarToolBar } from 'pages/scheduler/components/Toolbar';
+import { parseMeetingTimes, ParseMeetingTimesResult } from 'pages/scheduler/shared/parsers';
 import { CustomEvent } from 'pages/scheduler/shared/types';
 import { eventPropGetter } from 'pages/scheduler/styles/eventPropGetter';
 import { slotPropGetter } from 'pages/scheduler/styles/slotPropGetter';
@@ -29,83 +29,6 @@ const localizer = dateFnsLocalizer({
   getDay,
   locales: { 'en-US': enUS },
 });
-
-const parseMeetingTimeDays = (calendarEvent: CourseCalendarEvent) => {
-  const days = calendarEvent.meetingTime.days;
-  const daysRRule: Weekday[] = [];
-
-  if (days.includes('M')) {
-    daysRRule.push(RRule.MO);
-  }
-  if (days.includes('T')) {
-    daysRRule.push(RRule.TU);
-  }
-  if (days.includes('W')) {
-    daysRRule.push(RRule.WE);
-  }
-  if (days.includes('R')) {
-    daysRRule.push(RRule.TH);
-  }
-  if (days.includes('F')) {
-    daysRRule.push(RRule.FR);
-  }
-
-  return daysRRule;
-};
-
-const parseFormat = 'MM, d, yyyy h:mm a XXX';
-
-type ParseMeetingTimesResult = {
-  lower: RRule;
-  upper: RRule;
-  startDate: Date;
-  endDate: Date;
-};
-
-// TODO: try to move this into the backend as much as possible
-const parseMeetingTimes = (event: CourseCalendarEvent): ParseMeetingTimesResult => {
-  const lowerBound = parse(event.term, 'yyyyMM', new Date());
-
-  const startEndDates = event.meetingTime.dateRange.split('-').map((d) => d.replace(',', ''));
-  const startEndTimes = event.meetingTime.time.split('-').map((d) => d.trim());
-
-  // TODO: find better means of handling timezones
-  const courseStartDate = new Date(startEndDates[0] + ' 00:00:00 GMT');
-  const courseEndDate = new Date(startEndDates[1] + ' 00:00:00 GMT');
-
-  // getMonth returns 0-11, so we need to add 1 to get the correct month
-  const startDateString = `${lowerBound.getMonth() + 1}, 1, ${lowerBound.getUTCFullYear()}`;
-
-  const startUpperDateRRule = parse(`${startDateString} ${startEndTimes[0]} +00:00`, parseFormat, new Date());
-  const startLowerDateRRule = parse(`${startDateString} ${startEndTimes[1]} +00:00`, parseFormat, new Date());
-
-  const days = parseMeetingTimeDays(event);
-
-  // HACK: something doesn't like when start & end dates are the same
-  // adding 1 day to the end date makes everything happy :-)
-  courseEndDate.setDate(courseEndDate.getDate() + 1);
-
-  const ruleUpper = new RRule({
-    freq: RRule.WEEKLY,
-    byweekday: days,
-    dtstart: startUpperDateRRule,
-    until: courseEndDate,
-  });
-
-  const ruleLower = new RRule({
-    freq: RRule.WEEKLY,
-    byweekday: days,
-    dtstart: startLowerDateRRule,
-    until: courseEndDate,
-  });
-
-  return {
-    startDate: courseStartDate,
-    endDate: courseEndDate,
-    lower: ruleLower,
-    upper: ruleUpper,
-  };
-};
 
 export interface SchedulerCalendarProps {
   /**
